fix saving drafts
this is an addition to #1192:

by calling processComposeControls() twice, in onPause() and on Destroy(),
the seccond call gets an empty draft.

as onDestroy() is not guarateed to be executed anyway -
this was the intent of the move in pr #1192 -
we just remove this call and rely on onPause().

see https://developer.android.com/reference/android/app/Activity.html
for lifecycles and killable events.