stop threads and call close() on switching account
the imap-/mvbox-/sentbox-/smtp-threads needs to be terminated
before dc_close() can be called successfully.

if this is not done, they are hanging around, and, even worse,
if an account is opened _again_, there are now two threads
working on the database which lead to duplicated messages.

note, that this issue is only related to duplicated messages
if ever an account was selected, which, however, might be the same as the current.