fix marking incoming message as being seen
before, messages were only maked as seen after the chat was entered
or after some scrolling takes place.

if a user was inside a chat, reading the last incoming message,
that did not sent out markseen (unless, some scrolling was done,
which, honestly, is often the case)

this commit starts the markseen-bouncer also when the chatlist is reloaded
(true eg. on incoming messages) - but this is not done when the
list is paused (eg. screen is off or some other fragment/dialog above)