You won't believe how this makes the compilation 4x faster!
Sorry for the click-bait title, but it fitted too well ;-)

Anyway, I added the possibility to ndk-make.sh to compile only for one
arch.

In my tests, if I also set the optimisation level to 0 (default is 3),
the compilation time went down from 3:30 to 1:20. Not sure if this is
worth it because I remember a comment stating that this makes generating
keys take forever. Maybe I should do some measurements and experiment
with levels 1 and 2 but that's for another PR.

What do you think?