improve layout of emoji popups
in light mode, input field and emoji popups are both white,
therefore opening the emoji popup (used eg. for diversified emojis)
in the first line looks weird, you do not really know
if the display is part of the emoji picker or of the input field.
fixing that by adding a small border around the popup.

in dark mode, things are already fine.