Don't use quickContact in the ConversationListItem
quickContact means that you get to the contact details screen by
clicking the avatar. In the conversations list this is usually disabled,
just for the search it stayed enabled, which was unexpected
(because it was usually disabled ;-) ).

Even worse, in the search clicking the avatar sometimes got me to the
wrong contact details. No idea why, but I think disabling is better
anyway.