tweak universal_overlay color for black background
the color 'universal_overlay' can be used as backgrounds for
semi-transparent overlays.

the color guarantees enough contrast to white/light things drawn on it,
while still letting the background shine through.

the color is the same for both, dark and lite theme
as it is also used to draw on customized backgrounds
(where we can have any background anyway)

before, the shaped formed with this color are not visible
on black background, this is fixed by this commit,
making the color even more universal.

this is an improvement for the image-editor buttons
as well as for the info-messages on arbitrary backgrounds.