Red and more visible error icon
IIRC some time ago we talked about making the error icon red and make it
an exclamation mark to make it more visible.

Today someone complained to me that they thought the "x" looks like one
can close something and that it shouldn't be green, so I went ahead and
made it red :) (it's also a little bit bigger now)

- the image file is a vector (xml) file now, previously it was a png file.
Might this be a problem?