Dark theme / DayNight: Add option to follow system
fix maybe #1748. Maybe not, because actually I don't understand why
issue #1748 even happened because we don't use the system background for
the bubble but an own one (see conversation_item_sent.xml:50). Probably
the system is somehow trying to forcefully create a dark mode.

So, let's see whether this PR fixes the issue, if not, I'll try again.