Bugfix: open images in the media preview again
Since #1834, images were not opened in the media preview again but
opened with an external app.

This was because since #1834, only slides that return true on
`hasImage()` are shown in the media preview, i.e. gifs, images and
videos.

Now, the problem was that images were shown as DocumentSlides under
some circumstances, which return `false` for `hasImage()` (naturally).