refine wording of remaining screenlock
we keep using the screenlock for "Backup" and "Manage keys";
these dialog always used the wording of the "global" screenlock,
sth. as "Please Unlock Delta Chat", however, that wording was always wrong
as Delta Chat itself is not locked in these situations.

this pr sets the title to the selected action
and just says "Please enter system secret to continue."

moreover, the toasts on canceling the action are removed,
they do not make much sense here,
there was a user action just before, so that is pretty clear anyway.