consider multi-device usage on withdraw/revive QR code wordings
in a multi-device-setup,
if deviceB scans a setup-contact QR code shown on deviceA,
it is misleading to say that the QR code was withdrawn
and can be revived.
actually, the QR code was never withdrawn by the user.

this is fixed by this pr.
(at some point, we want to sync QR codes across devices,
but even then, there may be siutuations where things are not in sync)

moreover,
the wording is more explicit about what is usually done with the qr code;
this is esp. helpful, when scanning a printed QR code without much context.