capitalize some strings
as discussed several times on different channels,
nowadays, it is more usual to use capitalized titles;
fitting better to existing environments.

this pr capitalizes menu entries, buttons and titles;
"titles" include all name of all settings.

whats left lowercase (just to have a clear focus):
- states
- values
- whole sentences
- titles mentioned in whole sentences

(nb: there is the rough idea of having something as "Group or Subject"
instead of just "Group" -
this change is also easier with the new capitalisation rules ;)

as this all does not affect other languages
(they have their own rules),
we should be careful not to force re-translations.