remove 'Done' toast to avoid problems with lost activity
for whatever reason,
i just got a crash related to the toast that could not find its activity,
probably we should have used the app context and not the activity.

however, the toast was annoying me already,
there is no point in adding it, the change is visible instantly to the user
on the screen.