Suppress NewApi error
Android Studio showed an error (not just a warning) on
"ScribbleActivity.SCRIBBLE_REQUEST_CODE" because ScribbleActivity is
annotated as `@RequiresApi(19)`. Apparently there is no problem with
this because this has been this way for ages now and DeltaChat is
running fine on pre-API-19 devices, but it was really annoing that
Android Studio showed an error.