simpler account switching
remove the "progress" dialog that shows up for a fraction of a second.
as all accounts are already in memory since the switch to dc_accounts_t,
there is nothing that can take so much time to satisfy a "wait" dialog.

the switching is now at least visually nicer -
however, due to less task switching and less code,
it should also be a bit faster.