re-add "Add Attachment" menu entry
it was removed to have some less menu entries,
however, on recent discussions,
it turns out clearer and clearte that this will break some workflows
(writing text first, then adding attachment).

i just had a look at other "large" messengers -
all of them support "write text, then add attachment" -
as well as our iOS and Desktop apps.
also, this order seems to be pretty normal when it comes to "E-Mail".

all in all,
i think, we should not sacrify this workflow just to save one menu entry.

at some point,
maybe we want to have the "Attach" button always visible left of the input bar
as most other messengers do (as well as iOS and Desktop),
however, this is a larger thing and also has its drawbacks on smaller screens.

so, just leave things as is seems not to be the worse idea here.

the "View Profile" menu entry, however stays, removed,
so, there is some cleanup.