remove deprecated gradle element 'dexOptions'
without, there is a warning in gradle7 that says:

"DSL element 'dexOptions' is obsolete and should be removed.
It will be removed in version 8.0 of the Android Gradle plugin.
Using it has no effect, and the AndroidGradle plugin optimizes dexing automatically."

not sure, what the option was about in older gradles.