Fix unread count: Always show exactly the info that core gives us (#2164)
Fix #2163 by making the unread badge visible in the chatlist again. The
user can then go into this chat once in order to mark it as noticed.

Since https://github.com/deltachat/deltachat-core-rust/pull/2861,
the core will prevent the situation "the last message
is outgoing but there are fresh messages" from newly coming up, by
automatically marking the messages as noticed.

This PR removes all the logic that involved some "if the last message is
not fresh|noticed, the unread count should be 0" as this kind of logic
belongs to the core, because otherwise 1) have different behavior on
different platforms 2) issues like #2163 appear.

Now, if the _unread count_ is non-zero, the _unread badge_ is shown.

If the state of the _last message_ is "fresh", the line with the _last
message_ (called "subject line") is shown in bold.