add "Resend" to message menu
the UI does not check before selection if a message can be resent,
this is subject to change and not worth the effort.
instead, just an error is shown.

for the ID_NEVER_SHOW_AS_ACTION_* - this is a bit ugly,
however, the few alternatives i tried were even more uglier,
Java and Array-primitives are just not great.
as it is not changed often, it is probably not worth the effort to dive deeper.