use less muted color for "recently seen" (#2393)
with the "connection view",
we already have a green indicating sth. as "online", reuse that.
(the green used before is a bit too much of a muted color,
this does not reflect "activity" well)

i came over that when playing around with the layout on ios;
ios uses the same color now.

also many other apps use a similar, less muted green for "activity" or
"online".