fix crash in share-to-delta
the crash was introduced by 1.34.2 that was released for testing only.

reason is a wrong `exported` flag - it makes sense to have `exported="true"` here.

the flag was added in https://github.com/deltachat/deltachat-android/pull/2407
as needed for android12 - and i was assuming that the old default was `false`
and set that value to all missing ones.

that assumption seems to be wrong, so we should double-check all flags
changed by https://github.com/deltachat/deltachat-android/pull/2407