add second device first screen: open fragment instead of activity
idea was to add an image, a headline, a text, a button and a link to help.

however, on very short tries,
it seems not "just so" possible to add that to a preference fragment.

- we would need another activity: this would be more complicated as
  we have another "thing" in the already "special" flow
  (wrt keeping app active and restartable, we remove and restore back stack,
  see recent PRs)

- or, that would be my preferred solution,
  check how buttons etc. are possible within a preference fragment
  or, maybe we can add a non-preference fragment?
  idea is to stay inside the preference activity, however

probably not too complicated,
but as this is not necessary for 1.36,
a bit postponed.
also, we do not really have a nice image yet.
(we could see if sth. from https://undraw.co fits, eg. search for "device")