do not intercept requests when loading remote content is disabled
even with setBlockNetworkLoads(true),
shouldInterceptRequest() is called,
not sure if this is a bug or expected behavior,
i did not find documentation for that;
intuitively, i dit not expect sort of a proxy to bypass
and explicit block_networks = true,
however, one can probably also argue in the other direction.

leaving setBlockNetworkLoads(true),
in case default behavior changes (or older webviews have different ones),
we have a second layer.