reword 'Copy to Clipboard' to 'Copy Text'
message list allows multiple messages
being copied to clipboard at the same time;
if these messages as not just text, a summary is added,
sth. as "Voice Messages" or "Forwarded: Image - this is my family" etc.

this is on purpose and the gist of the action,
renaming the action to 'Copy Text' makes that more clear.