use system camera as default
the system camera has usually more features (flash, physical zoom, filter)
and is less buggy than the mostly unmaintained built-in one.
also, the system camera seems to be more supported
(isDeviceSupported() is not needed) and does not require to ask for permission.

there is still a switch in "advanced settings" to use the old built-in camera,
however, at some point we may remove the old code completely.