resume settings activity on language changes
this applies changed immediately,
so esp. right-to-left-changes are correct -
but also weird issues wrt wrong language in the title bar on some api should be fixed.

little downside is that after changing the language,
you're back to the "main settings page"
and do not stay in the "appearance page";
but you can regard this also as a feature, one tap less.

(reason is that the activity is resumed on language changes
and the underlying intent does not track fragments currently)