use 'Verified by me' instead 'Verified by you'
when talking about this function,
it is most times called 'Verified by me'.

this also makes translations easier to languages that
have diferent forms (as 'Du' vs. 'Sie' in german)

even though, we're most times using 'You' when the 'App talks to you',
we have some exceptions as 'Me' in the summary already,
where the 'You' does not work well.

(we leave the key as is, this is not really shown to translators anyways,
better key might be 'verified_by_self', however, it is not worth
opening this issue as the key is already in use -
and it is also not needed to force retranslation,
locales should use what fits and sounds best, may it be 'Me' or 'You')