always open account switcher from main title bar
avatar and title appears as one unit to the user;
doing different things on taps in the avatar (open account switcher)
and the title (open connectivity)
is confusing - and also easily done wrong.

this pr always opens the account switcher;
to get to the connetivity view,
there is a static button inside the account switchter.