make back button work in webxdc
this pr switches the "hardware" back button
from always terminating the app
to going back in history first,
only if the history is empty, the app is terminated.

this enables a webxdc to use document.history as additional navigation
as usual.

the "back" upper left of the screen still always exits the app.