revert 'introduced by' to 'verified by'
while it is still true that we need
to avoid the term 'verified' in the context of groups,
removing 'verified' completely at #2804 was a bit over the top.

for contacts, this term is still better than 'introduced'
which opens the quesions if not all contacts are somehow introduced.

so, to sum up, we stroke 'verified group' and 'verified checkmarks'
in favor to guaranteed 'end-to-end encryption' -
but we leave 'verified contact' as is.