make 1to1-empty-hint working better with chatmail
the hint already led to confusion on 37c3 testings of chatmail.

the hint that appears when a one-to-one-chat was just created manually
(by entering an email address manually) is a bit misleading when using chatmail,
as it says it 'does not matter if the other side uses Delta Chat'.

this is still true, however, with chatmail it is needed that messages
outside the chatmail-server are encrypted,
which is arguably not so easy without delta chat and QR code scanning.

this PR therefore just removes the hint;
it would still be great if the error message 'cannot encrypt' is improved,
however, this is another thing.