profile switcher: hide email for chatmail profiles
as on other places,
hide chatmail email address from primary UI as well in the account switcher.

non-chatmail addresses will stay for now;
we are considering to add a "tag" or sth. like that for the user
to differenciate between accounts if they're using same name+avatar.

this also makes the dialog more on-point and easier to read.