refresh the title bar color
after all the years, the blueish green
could need a freshup.

the new color is close enough,
that it does not break layout of other apps,
existing screenshots and so on.
that can be updated without stress as needed.

the new color was inspired by "Rinmans Green",
which has the additional safetly feature
to be known as an antibacterial color!