do not start foreground service in foreground
the foreground service was introduced unconditionally in #3312,
however turns out to be annoying and flickering
if the app is actually in foreground.

the service _might_ be needed in this case only
if the app is about going to background during fetch,
but in this case, we assume,
that we have some 10 seconds time still.

in any case,
this is still an improvement to the released apps,
that _never_ show a foreground service.