simplify e2ee strings
as discussed with @hpk42, we want to simplify the terms used in UI,
as there is no switch-forth-and-back, the "from now on" is superfluous,
and the "guaranteed" at least noisy and raises questions.

also, this allows us to use the same string when establishing via vcard contacts,
which is a good thing to move forward just now.

the faq will explain nuances then